requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
lxml>=5.0
